requirements.txt
discord.py
torch
transformers
aiohttp
pandas
python-dotenv

